+++ test/ustore/test_blob_node.cc
@@ -18,18 +18,18 @@ TEST(BlobNode, Basic) {
   EXPECT_TRUE(bnode.isLeaf());
   EXPECT_EQ(sizeof(raw_data), bnode.numEntries());
   EXPECT_EQ(sizeof(raw_data), bnode.numElements());
-  EXPECT_EQ(15, bnode.GetLength(10, 25));
-  EXPECT_EQ(1, bnode.len(10));
+  EXPECT_EQ(size_t(15), bnode.GetLength(10, 25));
+  EXPECT_EQ(size_t(1), bnode.len(10));
   EXPECT_EQ(0, std::memcmp(bnode.data(15), raw_data + 15, 1));
 
   ustore::byte_t* buffer = new ustore::byte_t[15];
   // Check normal copy
-  ASSERT_EQ(15, bnode.Copy(10, 15, buffer));
+  ASSERT_EQ(size_t(15), bnode.Copy(10, 15, buffer));
   ASSERT_EQ(0, std::memcmp(buffer, raw_data + 10, 15));
 
   // Check num_bytes to copy exceed the end of chunk
   size_t start_idx = sizeof(raw_data) - 5;
-  ASSERT_EQ(5, bnode.Copy(start_idx, 10, buffer));
+  ASSERT_EQ(size_t(5), bnode.Copy(start_idx, 10, buffer));
   ASSERT_EQ(0, std::memcmp(buffer, raw_data + start_idx, 5));
   delete[] buffer;
 }
@@ -46,12 +46,12 @@ TEST(BlobChunker, Basic) {
   const ustore::byte_t r[] = "aabbb";
   EXPECT_EQ(0, memcmp(chunk_info.chunk.data(), r, 5));
 
-  ASSERT_EQ(1, chunk_info.meta_seg->numEntries());
+  ASSERT_EQ(size_t(1), chunk_info.meta_seg->numEntries());
   const ustore::byte_t* me_data = chunk_info.meta_seg->entry(0);
   ustore::MetaEntry me(me_data);
 
   EXPECT_EQ(chunk_info.chunk.hash(), me.targetHash());
   EXPECT_EQ(chunk_info.meta_seg->numBytes(), me.numBytes());
-  EXPECT_EQ(5, me.numElements());
-  EXPECT_EQ(1, me.numLeaves());
+  EXPECT_EQ(size_t(5), me.numElements());
+  EXPECT_EQ(size_t(1), me.numLeaves());
 }
