+++ test/ustore/test_chunker.cc
@@ -10,8 +10,8 @@ TEST(FixedChunker, Basic) {
   ustore::FixedSegment seg1(data, 5, 1);
 
   EXPECT_EQ(*(data + 2), *(seg1.entry(2)));
-  EXPECT_EQ(1, seg1.entryNumBytes(1));
-  EXPECT_EQ(5, seg1.numBytes());
+  EXPECT_EQ(size_t(1), seg1.entryNumBytes(1));
+  EXPECT_EQ(size_t(5), seg1.numBytes());
   EXPECT_FALSE(seg1.empty());
   EXPECT_EQ(data, seg1.data());
 
@@ -22,9 +22,9 @@ TEST(FixedChunker, Basic) {
 
   ustore::FixedSegment seg2(data, 2);
   EXPECT_TRUE(seg2.empty());
-  ASSERT_EQ(1, seg2.prolong(2));
-  EXPECT_EQ(1, seg2.numEntries());
-  EXPECT_EQ(2, seg2.numBytes());
+  ASSERT_EQ(size_t(1), seg2.prolong(2));
+  EXPECT_EQ(size_t(1), seg2.numEntries());
+  EXPECT_EQ(size_t(2), seg2.numBytes());
 }
 
 TEST(FixedChunker, Split) {
@@ -34,24 +34,24 @@ TEST(FixedChunker, Split) {
   // Split at start
   auto splits = seg1.Split(0);
   EXPECT_TRUE(splits.first->empty());
-  EXPECT_EQ(5, splits.second->numBytes());
-  EXPECT_EQ(5, splits.second->numEntries());
+  EXPECT_EQ(size_t(5), splits.second->numBytes());
+  EXPECT_EQ(size_t(5), splits.second->numEntries());
   EXPECT_EQ(0, memcmp(data, splits.second->data(), 5));
 
   // Split at Middle
   splits = seg1.Split(2);
-  EXPECT_EQ(2, splits.first->numBytes());
-  EXPECT_EQ(2, splits.first->numEntries());
+  EXPECT_EQ(size_t(2), splits.first->numBytes());
+  EXPECT_EQ(size_t(2), splits.first->numEntries());
   EXPECT_EQ(0, memcmp(data, splits.first->data(), 2));
 
-  EXPECT_EQ(3, splits.second->numBytes());
-  EXPECT_EQ(3, splits.second->numEntries());
+  EXPECT_EQ(size_t(3), splits.second->numBytes());
+  EXPECT_EQ(size_t(3), splits.second->numEntries());
   EXPECT_EQ(0, memcmp(data + 2, splits.second->data(), 3));
 
   // Split at End
   splits = seg1.Split(5);
-  EXPECT_EQ(5, splits.first->numBytes());
-  EXPECT_EQ(5, splits.first->numEntries());
+  EXPECT_EQ(size_t(5), splits.first->numBytes());
+  EXPECT_EQ(size_t(5), splits.first->numEntries());
   EXPECT_EQ(0, memcmp(data, splits.first->data(), 5));
   EXPECT_TRUE(splits.second->empty());
 
@@ -63,9 +63,9 @@ TEST(VarChunker, Basic) {
   ustore::VarSegment seg1(data, 5, {0, 2});
 
   EXPECT_EQ(*(data + 2), *(seg1.entry(1)));
-  EXPECT_EQ(2, seg1.entryNumBytes(0));
-  EXPECT_EQ(3, seg1.entryNumBytes(1));
-  EXPECT_EQ(5, seg1.numBytes());
+  EXPECT_EQ(size_t(2), seg1.entryNumBytes(0));
+  EXPECT_EQ(size_t(3), seg1.entryNumBytes(1));
+  EXPECT_EQ(size_t(5), seg1.numBytes());
   EXPECT_FALSE(seg1.empty());
   EXPECT_EQ(data, seg1.data());
 
@@ -77,15 +77,15 @@ TEST(VarChunker, Basic) {
   ustore::VarSegment seg2(data);
   EXPECT_TRUE(seg2.empty());
 
-  ASSERT_EQ(1, seg2.prolong(2));
-  EXPECT_EQ(1, seg2.numEntries());
-  EXPECT_EQ(2, seg2.entryNumBytes(0));
-  EXPECT_EQ(2, seg2.numBytes());
+  ASSERT_EQ(size_t(1), seg2.prolong(2));
+  EXPECT_EQ(size_t(1), seg2.numEntries());
+  EXPECT_EQ(size_t(2), seg2.entryNumBytes(0));
+  EXPECT_EQ(size_t(2), seg2.numBytes());
 
-  ASSERT_EQ(2, seg2.prolong(1));
-  EXPECT_EQ(2, seg2.numEntries());
-  EXPECT_EQ(1, seg2.entryNumBytes(1));
-  EXPECT_EQ(3, seg2.numBytes());
+  ASSERT_EQ(size_t(2), seg2.prolong(1));
+  EXPECT_EQ(size_t(2), seg2.numEntries());
+  EXPECT_EQ(size_t(1), seg2.entryNumBytes(1));
+  EXPECT_EQ(size_t(3), seg2.numBytes());
 }
 
 TEST(VarChunker, Split) {
@@ -95,24 +95,24 @@ TEST(VarChunker, Split) {
   // Split at start
   auto splits = seg1.Split(0);
   EXPECT_TRUE(splits.first->empty());
-  EXPECT_EQ(5, splits.second->numBytes());
-  EXPECT_EQ(3, splits.second->numEntries());
+  EXPECT_EQ(size_t(5), splits.second->numBytes());
+  EXPECT_EQ(size_t(3), splits.second->numEntries());
   EXPECT_EQ(0, memcmp(data, splits.second->data(), 5));
 
   // Split at Middle
   splits = seg1.Split(2);
-  EXPECT_EQ(3, splits.first->numBytes());
-  EXPECT_EQ(2, splits.first->numEntries());
+  EXPECT_EQ(size_t(3), splits.first->numBytes());
+  EXPECT_EQ(size_t(2), splits.first->numEntries());
   EXPECT_EQ(0, memcmp(data, splits.first->data(), 3));
 
-  EXPECT_EQ(2, splits.second->numBytes());
-  EXPECT_EQ(1, splits.second->numEntries());
+  EXPECT_EQ(size_t(2), splits.second->numBytes());
+  EXPECT_EQ(size_t(1), splits.second->numEntries());
   EXPECT_EQ(0, memcmp(data + 3, splits.second->data(), 2));
 
   // Split at End
   splits = seg1.Split(3);
-  EXPECT_EQ(5, splits.first->numBytes());
-  EXPECT_EQ(3, splits.first->numEntries());
+  EXPECT_EQ(size_t(5), splits.first->numBytes());
+  EXPECT_EQ(size_t(3), splits.first->numEntries());
   EXPECT_EQ(0, memcmp(data, splits.first->data(), 5));
   EXPECT_TRUE(splits.second->empty());
