+++ test/ustore/test_http.cc
@@ -248,9 +248,9 @@ TEST(HttpTest, BasicOps) {
     workers.push_back(new ustore::WorkerService(worker_addr, "", false));
 
   std::vector<std::thread> worker_threads;
-  for (int i = 0; i < workers.size(); i++)
+  for (size_t i = 0; i < workers.size(); i++)
     workers[i]->Init();
-  for (int i = 0; i < workers.size(); i++)
+  for (size_t i = 0; i < workers.size(); i++)
     worker_threads.push_back(std::thread(&WorkerService::Start, workers[i]));
 
   // launch clients
@@ -415,7 +415,7 @@ TEST(HttpTest, BasicOps) {
   client_service_thread.join();
   usleep(kSleepTime);
   // stop workers
-  for (int i = 0; i < worker_threads.size(); i++) {
+  for (size_t i = 0; i < worker_threads.size(); i++) {
     workers[i]->Stop();
     worker_threads[i].join();
     delete workers[i];
