+++ test/ustore/test_list_node.cc
@@ -30,20 +30,20 @@ class ListNodeEnv : public ::testing::Test {
 };
 
 TEST_F(ListNodeEnv, Chunker) {
-  ASSERT_EQ(1, chunk_info_.meta_seg->numEntries());
+  ASSERT_EQ(size_t(1), chunk_info_.meta_seg->numEntries());
   const ustore::byte_t* me_data = chunk_info_.meta_seg->entry(0);
   ustore::MetaEntry me(me_data);
 
   EXPECT_EQ(chunk_info_.chunk.hash(), me.targetHash());
   EXPECT_EQ(chunk_info_.meta_seg->numBytes(), me.numBytes());
-  EXPECT_EQ(3, me.numElements());
-  EXPECT_EQ(1, me.numLeaves());
+  EXPECT_EQ(uint64_t(3), me.numElements());
+  EXPECT_EQ(uint32_t(1), me.numLeaves());
 }
 
 TEST_F(ListNodeEnv, ListNode) {
   ustore::ListNode lnode(&chunk_info_.chunk);
-  EXPECT_EQ(3, lnode.numEntries());
-  EXPECT_EQ(3, lnode.numElements());
+  EXPECT_EQ(size_t(3), lnode.numEntries());
+  EXPECT_EQ(uint64_t(3), lnode.numElements());
   const ustore::Slice actual_element2 = ustore::ListNode::Decode(lnode.data(1));
   EXPECT_EQ(element2.len(), actual_element2.len());
   EXPECT_EQ(0,
